requirements.txt
click==8.1.3
Flask==2.2.1
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
importlib-metadata==4.12.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pyactiveresource==2.2.2
PyJWT==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
PyYAML==6.0
ShopifyAPI==12.0.0
six==1.16.0
Werkzeug==2.2.1
zipp==3.8.1
